refactor: address PR review feedback
